moves Dynatrace API DTOs to own file
